requirements.txt
Flask==3.0.3
gunicorn==22.0.0
tensorflow==2.16.1
numpy==1.26.4
scikit-learn==1.5.0
Pillow==10.3.0
opencv-python-headless==4.9.0.80
matplotlib==3.8.4
seaborn==0.13.2
Werkzeug==3.0.3